Rename the Reader and Writer counter interfaces.
